+++ R/imports.R
@@ -1,5 +1,5 @@
-#' @importFrom utils browseURL
 #' @importFrom stats na.omit aggregate
-#' @importFrom stringr str_trim
+#' @importFrom stringr fixed regex str_count str_match_all str_trim 
 #' @importFrom yamlme write_rmd
+#' @importFrom methods setOldClass
 NULL